Factor workspaces listing out of utils into appropriate tile
