+++ src/ploneintranet/workspace/utils.py
@@ -4,10 +4,10 @@ from BTrees.OOBTree import OOBTree
 from ploneintranet.workspace.workspacefolder import IWorkspaceFolder
 from Products.CMFCore.interfaces import ISiteRoot
 from plone import api
-from ploneintranet.microblog.interfaces import IMicroblogTool
+
 from zope.annotation import IAnnotations
 from zope.component import getUtility
-from zope.component import queryUtility
+
 from collective.workspace.interfaces import IWorkspace
 from ploneintranet.workspace import MessageFactory as _
 
@@ -75,64 +75,6 @@ def parent_workspace(context):
 
 def in_workspace(context):
     return IWorkspaceFolder.providedBy(parent_workspace(context))
-
-
-def escape_id_to_class(cid):
-    """ We use workspace ids as classes to style them.
-        if a workspace has dots in its name, this is not usable as a class
-        name. We have to escape that. We might need to do more to them, so this
-        became a utility function.
-    """
-    return cid.replace('.', '-')
-
-
-def get_workspace_activities(brain, limit=1):
-    """ Return the workspace activities sorted by reverse chronological
-    order
-
-    Regarding the time value:
-     - the datetime value contains the time in international format
-       (machine readable)
-     - the title value contains the absolute date and time of the post
-    """
-    mb = queryUtility(IMicroblogTool)
-    items = mb.context_values(brain.getObject(), limit=limit)
-    mtool = api.portal.get_tool('portal_membership')
-    results = []
-    for item in items:
-        user_data = mtool.getMemberInfo(item.creator)
-        creator = user_data.get('fullname') if user_data else item.creator
-        results.append(dict(
-            subject=creator,
-            verb='published',
-            object=item.text,
-            time={
-                'datetime': item.date.strftime('%Y-%m-%d'),
-                'title': item.date.strftime('%d %B %Y, %H:%M')}
-        ))
-    return results
-
-
-def my_workspaces(context):
-    """ The list of my workspaces
-    """
-    pc = api.portal.get_tool('portal_catalog')
-    brains = pc(
-        portal_type="ploneintranet.workspace.workspacefolder",
-        sort_on="modified",
-        sort_order="reversed",
-    )
-    workspaces = [
-        {
-            'id': brain.getId,
-            'title': brain.Title,
-            'description': brain.Description,
-            'url': brain.getURL(),
-            'activities': get_workspace_activities(brain),
-            'class': escape_id_to_class(brain.getId),
-        } for brain in brains
-    ]
-    return workspaces
 
 
 def existing_users(context):
